Fabric: set epoch for batch_sampler.sampler too
